final clipping ui
