+++ packages/app/app/studio/[organization]/layout.tsx
@@ -37,7 +37,7 @@ const Layout = async ({
           organizations={userData?.organizations}
         />
 
-        <div className="top-[64px] overflow-hidden flex flex-col h-[calc(100vh-64px)] border-t border-secondary">
+        <div className="top-[54px] overflow-hidden flex flex-col h-[calc(100vh-54px)] border-t border-secondary">
           {!hasOrganization(
             userData?.organizations,
             params.organization
@@ -58,7 +58,7 @@ const Layout = async ({
             </div>
           ) : (
             <div className="flex flex-row w-full h-full">
-              <div className="overflow-y-auto w-full h-full">
+              <div className="w-full h-full">
                 {children}
               </div>
             </div>
