+++ packages/app/components/ui/navigation-menu.tsx
@@ -12,7 +12,7 @@ const NavigationMenu = React.forwardRef<
   <NavigationMenuPrimitive.Root
     ref={ref}
     className={cn(
-      'relative z-10 flex w-full items-center justify-center',
+      'relative  flex w-full items-center justify-center',
       className
     )}
     {...props}>
